refactor: 分析モジュールの責務分離とアイコンの絵文字化
- personality_analyzer.dartとenhanced_personality_analyzer.dartを小さなモジュールに分割
- 各サービスが単一責任を持つように責務を分離
- カードアイコンをIconウィジェットから絵文字に変更
- ファイル名から_refactoredサフィックスを削除
