+++ lib/features/analysis/analysis.dart
@@ -1,0 +1,19 @@
+/// 分析機能のバレルファイル
+/// このファイルをインポートするだけで、必要な分析機能にアクセスできます
+
+// モデル
+export 'models/analysis_result.dart';
+export 'models/personality_type.dart';
+
+// 定数
+export 'constants/personality_constants.dart';
+
+// サービス
+export 'services/score_calculator.dart';
+export 'services/personality_determiner.dart';
+
+// 洞察生成
+export 'insights/insight_generator.dart';
+
+// メインのアナライザー
+export 'personality_analyzer.dart';
