+++ lib/features/analysis/enhanced_analysis.dart
@@ -1,0 +1,16 @@
+/// 拡張分析機能のバレルファイル
+/// このファイルをインポートするだけで、必要な拡張分析機能にアクセスできます
+
+// モデル
+export 'models/enhanced_models.dart';
+
+// サービス
+export 'services/motivation_analyzer.dart';
+export 'services/growth_stage_analyzer.dart';
+export 'services/time_series_analyzer.dart';
+
+// 洞察生成
+export 'insights/combined_insight_generator.dart';
+
+// メインのアナライザー
+export 'enhanced_personality_analyzer.dart';
